Refactor evaluator to be outside the checker.
