+++ tests/baselines/reference/isolated-declarations/auto-fixed/tsc/constEnum2.d.ts
@@ -1,0 +1,57 @@
+//// [tests/cases/conformance/constEnums/constEnum2.ts] ////
+
+//// [constEnum2.ts]
+// An enum declaration that specifies a const modifier is a constant enum declaration.
+// In a constant enum declaration, all members must have constant values and
+// it is an error for a member declaration to specify an expression that isn't classified as a constant enum expression.
+
+// Error : not a constant enum expression
+
+const CONST: number = 9000 % 2;
+const enum D {
+    d = 10,
+    e = 199 * Math.floor(Math.random() * 1000),
+    f = d - (100 * Math.floor(Math.random() % 8)),
+    g = CONST,
+}
+
+/// [Declarations] ////
+
+
+
+//// [constEnum2.d.ts]
+declare const CONST: number;
+declare const enum D {
+    d = 10,
+    e,
+    f,
+    g
+}
+//# sourceMappingURL=constEnum2.d.ts.map
+/// [Errors] ////
+
+constEnum2.ts(10,9): error TS2474: const enum member initializers must be constant expressions.
+constEnum2.ts(11,9): error TS2474: const enum member initializers must be constant expressions.
+constEnum2.ts(12,9): error TS2474: const enum member initializers must be constant expressions.
+
+
+==== constEnum2.ts (3 errors) ====
+    // An enum declaration that specifies a const modifier is a constant enum declaration.
+    // In a constant enum declaration, all members must have constant values and
+    // it is an error for a member declaration to specify an expression that isn't classified as a constant enum expression.
+    
+    // Error : not a constant enum expression
+    
+    const CONST: number = 9000 % 2;
+    const enum D {
+        d = 10,
+        e = 199 * Math.floor(Math.random() * 1000),
+            ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+!!! error TS2474: const enum member initializers must be constant expressions.
+        f = d - (100 * Math.floor(Math.random() % 8)),
+            ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+!!! error TS2474: const enum member initializers must be constant expressions.
+        g = CONST,
+            ~~~~~
+!!! error TS2474: const enum member initializers must be constant expressions.
+    }
